requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
emoji==0.5.4
future==0.18.2
geographiclib==1.50
geopy==2.0.0
idna==2.9
pycparser==2.20
python-telegram-bot==12.7
requests==2.24.0
six==1.14.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.9
